Kahans Algorithm
Using Kahan's algorithm for compensated summation, as it takes
4 addition/subtractions per update, instead of the Welford
method, which is 2 add/sub, and 2 mult/div. Moreover, by
couching the problem as summation, we easily add and subtract
an element simultaneously, so the robust kabuki needs only be
done once per output. The end result is 5 add/subs per
update on a rolling window, instead of 4 add/subs and 4 mult/div,
approximately.
This gives good speedups for both running_sum and running_mean.