another reference timing issue
so I think using swap properly is the last remaining hurdle here
part of the issue is the random junk of checking bounds,
and part is the use of add and remove separately instead
of a proper swap.
not sure how this can be made faster, other than having
bespoke code for running_sd.
note that the swap code should be expanded to the weighted case.